+++ src/main.ts
@@ -1,0 +1,7 @@
+import { logger } from './logger';
+import { env } from './environment';
+
+async function main() {
+	logger.info('Hello ' + env.APP_TITLE);
+}
+main();